Added QFT doc comments, tests
