+++ library/std/canon.qs
@@ -476,15 +476,50 @@ namespace Microsoft.Quantum.Canon {
         }
     }
 
-    // TODO: This needs description.
-    // TODO: Need to figure out if it's actually "LE".
-    operation QFTLE (qs : Qubit[]) : Unit is Adj {
-        open Microsoft.Quantum.Arrays;
-        for (i, q) in Reversed(Enumerated(qs)) {
-            H(q);
-            for (j, q2) in Enumerated(Reversed(qs[...i - 1])) {
-                Controlled R1Frac([q], (1, j + 1, q2));
+    /// # Summary
+    /// Applies Quantum Fourier Transform (QFT) to a quantum register.
+    ///
+    /// # Description
+    /// Applies QFT to a little-endian register `qs` of length n
+    /// containing |x₁⟩⊗|x₂⟩⊗…⊗|xₙ⟩. The qs[0] contains the
+    /// least significant bit xₙ. The state of qs[0] becomes
+    /// (|0⟩+𝑒^(2π𝑖[0.xₙ])|1⟩)/sqrt(2) after the operation.
+    ///
+    /// # Input
+    /// ## qs
+    /// Quantum register in a little-endian format to which the QFT is applied.
+    ///
+    /// # Reference
+    ///  - [Quantum Fourier transform](https://en.wikipedia.org/wiki/Quantum_Fourier_transform)
+    operation ApplyQFT (qs : Qubit[]) : Unit is Adj + Ctl {
+        let length = Length(qs);
+        Fact(length >= 1, "ApplyQFT: Length(qs) must be at least 1.");
+        for i in length-1..-1..0 {
+            H(qs[i]);
+            for j in 0..i-1 {
+                Controlled R1Frac([qs[i]], (1, j+1, qs[i-j-1]));
             }
+        }
+    }
+
+    /// # Summary
+    /// Applies Quantum Fourier Transform (QFT) to a quantum register
+    /// and reverses the result to match the big-endian format.
+    operation QFTLE (qs : Qubit[]) : Unit is Adj + Ctl {
+        ApplyQFT(qs);
+        SwapReverseRegister(qs);
+    }
+
+    /// # Summary
+    /// Uses SWAP gates to reverse the order of the qubits in a register.
+    ///
+    /// # Input
+    /// ## register
+    /// The qubits order of which should be reversed using SWAP gates
+    operation SwapReverseRegister (register : Qubit[]) : Unit is Adj + Ctl {
+        let length = Length(register);
+        for i in 0 .. length/2 - 1 {
+            SWAP(register[i], register[(length - i) - 1]);
         }
     }
