+++ library/tests/src/lib.rs
@@ -9,6 +9,8 @@ mod test_arithmetic;
 #[cfg(test)]
 mod test_arrays;
 #[cfg(test)]
+mod test_canon;
+#[cfg(test)]
 mod test_convert;
 #[cfg(test)]
 mod test_math;
